+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+#### 0.1.1.wip - MAU `2016-12-10`  
+**Added and Array, Date classes and custom assertions**
+
+* Added All Array methods from old gem
+* Added All Date methods from old gem
+    * Some needed to be rewritten and/or deprecated
+* Added <b>deprecate_this</b> for custom deprecation warnings
+* Added custom assertions and expectation for minitest
+* All methods should be fully documented to this point
+
+---
+
 #### 0.1.0.wip - MAU `2016-12-10`  
 **Initial Version**